remove data code
